move ssvc decision points into namespace-based directory
