+++ docs/reference/decision_points/exploitation.md
@@ -1,7 +1,7 @@
 # Exploitation
 
 ```python exec="true" idprefix=""
-from ssvc.decision_points.exploitation import LATEST
+from ssvc.decision_points.ssvc_.exploitation import LATEST
 from ssvc.doc_helpers import example_block
 
 print(example_block(LATEST))
@@ -49,7 +49,7 @@ The table below lists CWE-IDs that could be used to mark a vulnerability as *PoC
 ## Prior Versions
 
 ```python exec="true" idprefix=""
-from ssvc.decision_points.exploitation import VERSIONS
+from ssvc.decision_points.ssvc_.exploitation import VERSIONS
 from ssvc.doc_helpers import prior_version, example_block
 
 versions = VERSIONS[:-1]
