+++ docs/reference/decision_points/public_value_added.md
@@ -1,7 +1,7 @@
 # Public Value Added
 
 ```python exec="true" idprefix=""
-from ssvc.decision_points.public_value_added import LATEST
+from ssvc.decision_points.ssvc_.public_value_added import LATEST
 from ssvc.doc_helpers import example_block
 
 print(example_block(LATEST))
