+++ docs/reference/decision_points/report_public.md
@@ -1,7 +1,7 @@
 # Report Public
 
 ```python exec="true" idprefix=""
-from ssvc.decision_points.report_public import LATEST
+from ssvc.decision_points.ssvc_.report_public import LATEST
 from ssvc.doc_helpers import example_block
 
 print(example_block(LATEST))
