+++ docs/reference/decision_points/safety_impact.md
@@ -1,7 +1,7 @@
 # Safety Impact
 
 ```python exec="true" idprefix=""
-from ssvc.decision_points.safety_impact import LATEST
+from ssvc.decision_points.ssvc_.safety_impact import LATEST
 from ssvc.doc_helpers import example_block
 
 print(example_block(LATEST))
@@ -217,7 +217,7 @@ We defer this topic for now because we combine it with [*Mission Impact*](missio
 ## Prior Versions
 
 ```python exec="true" idprefix=""
-from ssvc.decision_points.safety_impact import VERSIONS
+from ssvc.decision_points.ssvc_.safety_impact import VERSIONS
 from ssvc.doc_helpers import example_block
 
 versions = VERSIONS[:-1]
