+++ docs/reference/decision_points/supplier_cardinality.md
@@ -1,7 +1,7 @@
 # Supplier Cardinality
 
 ```python exec="true" idprefix=""
-from ssvc.decision_points.supplier_cardinality import LATEST
+from ssvc.decision_points.ssvc_.supplier_cardinality import LATEST
 from ssvc.doc_helpers import example_block
 
 print(example_block(LATEST))
