+++ docs/reference/decision_points/supplier_involvement.md
@@ -1,7 +1,7 @@
 # Supplier Involvement
 
 ```python exec="true" idprefix=""
-from ssvc.decision_points.supplier_involvement import LATEST
+from ssvc.decision_points.ssvc_.supplier_involvement import LATEST
 from ssvc.doc_helpers import example_block
 
 print(example_block(LATEST))
