+++ docs/reference/decision_points/system_exposure.md
@@ -1,7 +1,7 @@
 # System Exposure
 
 ```python exec="true" idprefix=""
-from ssvc.decision_points.system_exposure import LATEST
+from ssvc.decision_points.ssvc_.system_exposure import LATEST
 from ssvc.doc_helpers import example_block
 
 print(example_block(LATEST))
@@ -44,7 +44,7 @@ If you have suggestions for further heuristics, or potential counterexamples to 
 ## Prior Versions
 
 ```python exec="true" idprefix=""
-from ssvc.decision_points.system_exposure import VERSIONS
+from ssvc.decision_points.ssvc_.system_exposure import VERSIONS
 from ssvc.doc_helpers import example_block
 
 versions = VERSIONS[:-1]
