+++ docs/reference/decision_points/technical_impact.md
@@ -1,7 +1,7 @@
 # Technical Impact
 
 ```python exec="true" idprefix=""
-from ssvc.decision_points.technical_impact import LATEST
+from ssvc.decision_points.ssvc_.technical_impact import LATEST
 from ssvc.doc_helpers import example_block
 
 print(example_block(LATEST))
