+++ docs/reference/decision_points/value_density.md
@@ -1,7 +1,7 @@
 # Value Density (SSVC)
 
 ```python exec="true" idprefix=""
-from ssvc.decision_points.value_density import LATEST
+from ssvc.decision_points.ssvc_.value_density import LATEST
 from ssvc.doc_helpers import example_block
 
 print(example_block(LATEST))
