+++ src/ssvc/decision_points/ssvc_/automatable.py
