+++ src/ssvc/decision_points/ssvc_/high_value_asset.py
