+++ src/ssvc/decision_points/ssvc_/safety_impact.py
