+++ src/ssvc/decision_points/ssvc_/technical_impact.py
