+++ src/ssvc/dp_groups/ssvc/deployer.py
@@ -18,20 +18,20 @@ created_at: 9/21/23 11:40 AM
 #  Carnegie Mellon®, CERT® and CERT Coordination Center® are registered in the
 #  U.S. Patent and Trademark Office by Carnegie Mellon University
 
-from ssvc.decision_points.automatable import AUTOMATABLE_2
-from ssvc.decision_points.exploitation import EXPLOITATION_1
-from ssvc.decision_points.human_impact import HUMAN_IMPACT_2
-from ssvc.decision_points.mission_impact import (
+from ssvc.decision_points.ssvc_.automatable import AUTOMATABLE_2
+from ssvc.decision_points.ssvc_.exploitation import EXPLOITATION_1
+from ssvc.decision_points.ssvc_.human_impact import HUMAN_IMPACT_2
+from ssvc.decision_points.ssvc_.mission_impact import (
     MISSION_IMPACT_1,
     MISSION_IMPACT_2,
 )
-from ssvc.decision_points.safety_impact import SAFETY_IMPACT_1
-from ssvc.decision_points.system_exposure import (
+from ssvc.decision_points.ssvc_.safety_impact import SAFETY_IMPACT_1
+from ssvc.decision_points.ssvc_.system_exposure import (
     SYSTEM_EXPOSURE_1,
     SYSTEM_EXPOSURE_1_0_1,
 )
-from ssvc.decision_points.utility import UTILITY_1_0_1
-from ssvc.decision_points.value_density import VALUE_DENSITY_1
+from ssvc.decision_points.ssvc_.utility import UTILITY_1_0_1
+from ssvc.decision_points.ssvc_.value_density import VALUE_DENSITY_1
 from ssvc.dp_groups.base import SsvcDecisionPointGroup
 
 PATCH_APPLIER_1 = SsvcDecisionPointGroup(
